Handshake Initial Commit
I wanted to use knowledge from the TOR handshake to make a new handshake for TOU.

but, in order to add another feature (auto updating the mod like TOR does), the handshake has to more closely resemble TORs one, so im disappointed to say i changed less than i wanted, but made a few improvements, and put a placeholder in for where this can link into the auto updating of the mod (if that will even work).

Either way, this fixes the crash bug, and ive left in the lobby features that were in the TOR handshake, because why remove them? i have hardcoded some to just work properly so settings can be coded in for them at a later date, and for other things ive implemented them as is as the features are just helpful.

Credit should go to TOR for the handshake template/design as well as the additional features, i have modified it a bit and thrown in a few extras/improvements, otherwiise it pretty much follows hows theirs works.